# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 62343467c97c70218d65043f
  title: couseModulePrototype
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  /module:
    post:
      x-skaffolder-id: 623438d5c97c70218d6504dd
      x-skaffolder-name: create
      x-skaffolder-id-resource: 62343732c97c70218d6504c5
      x-skaffolder-resource: Module
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - Module
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mDescription:
                  type: string
                mName:
                  type: string
    get:
      x-skaffolder-id: 623438ccc97c70218d6504da
      x-skaffolder-name: list
      x-skaffolder-id-resource: 62343732c97c70218d6504c5
      x-skaffolder-resource: Module
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - Module
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
  "/module/{id}":
    delete:
      x-skaffolder-id: 623438ccc97c70218d6504d8
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 62343732c97c70218d6504c5
      x-skaffolder-resource: Module
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - Module
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 623438d5c97c70218d6504de
      x-skaffolder-name: get
      x-skaffolder-id-resource: 62343732c97c70218d6504c5
      x-skaffolder-resource: Module
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - Module
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    post:
      x-skaffolder-id: 623438d5c97c70218d6504db
      x-skaffolder-name: update
      x-skaffolder-id-resource: 62343732c97c70218d6504c5
      x-skaffolder-resource: Module
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - Module
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Module"
...
  /user:
    post:
      x-skaffolder-id: 62343468c97c70218d65044d
      x-skaffolder-name: create
      x-skaffolder-id-resource: 62343468c97c70218d650449
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 62343468c97c70218d650454
      x-skaffolder-name: list
      x-skaffolder-id-resource: 62343468c97c70218d650449
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  "/user/{id}":
    delete:
      x-skaffolder-id: 62343468c97c70218d650450
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 62343468c97c70218d650449
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 62343468c97c70218d650452
      x-skaffolder-name: get
      x-skaffolder-id-resource: 62343468c97c70218d650449
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 62343468c97c70218d65044e
      x-skaffolder-name: update
      x-skaffolder-id-resource: 62343468c97c70218d650449
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Module:
      x-skaffolder-id: 62343732c97c70218d6504c5
      x-skaffolder-id-db: 62343467c97c70218d650440
      x-skaffolder-id-entity: 62343732c97c70218d6504c2
      x-skaffolder-url: /module
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mDescription:
          type: string
          x-skaffolder-id-attr: 62343732c97c70218d6504c4
          x-skaffolder-type: String
          x-skaffolder-required: true
        mName:
          type: string
          x-skaffolder-id-attr: 62343732c97c70218d6504c3
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - mDescription
        - mName
    User:
      x-skaffolder-id: 62343468c97c70218d650449
      x-skaffolder-id-db: 62343467c97c70218d650440
      x-skaffolder-id-entity: 62343467c97c70218d650441
      x-skaffolder-url: /user
      x-skaffolder-type: User
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 62343467c97c70218d650444
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 62343467c97c70218d650445
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 62343467c97c70218d650443
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 62343467c97c70218d650447
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 62343467c97c70218d650446
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 62343467c97c70218d650442
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
...