# src/main/java/com/cousemoduleprototype/db/cousemoduleprototype_db/entity/Module.java
package com.cousemoduleprototype.db.cousemoduleprototype_db.entity;

import com.cousemoduleprototype.db.cousemoduleprototype_db.entity.base.ModuleBase;

public class Module extends ModuleBase {

	//OVERRIDE HERE YOUR CUSTOM MAPPER
	
	
}